combining loader and result together
