+++ src/app/Scenes/ArtQuiz/ArtQuiz.tsx
@@ -1,17 +1,15 @@
 import { NavigationContainer } from "@react-navigation/native"
 import { createStackNavigator, TransitionPresets } from "@react-navigation/stack"
-import {
-  ArtQuizResultsNavigation,
-  ArtQuizResultsNavigationScreens,
-} from "app/Scenes/ArtQuiz/ArtQuizResults/navigation"
-import { ArtQuizArtworks } from "./ArtQuizArtworks"
+import { ArtQuizArtworks } from "app/Scenes/ArtQuiz/ArtQuizArtworks"
+import { ArtQuizResults } from "app/Scenes/ArtQuiz/ArtQuizResults/ArtQuizResults"
 import { ArtQuizWelcome } from "./ArtQuizWelcome"
 
 // tslint:disable-next-line:interface-over-type-literal
 export type ArtQuizNavigationStack = {
   ArtQuizWelcome: undefined
   ArtQuizArtworks: undefined
-} & ArtQuizResultsNavigationScreens
+  ArtQuizResults: undefined
+}
 
 export const StackNavigator = createStackNavigator<ArtQuizNavigationStack>()
 
@@ -28,7 +26,7 @@ export const ArtQuiz: React.FC = () => {
       >
         <StackNavigator.Screen name="ArtQuizWelcome" component={ArtQuizWelcome} />
         <StackNavigator.Screen name="ArtQuizArtworks" component={ArtQuizArtworks} />
-        {ArtQuizResultsNavigation()}
+        <StackNavigator.Screen name="ArtQuizResults" component={ArtQuizResults} />
       </StackNavigator.Navigator>
     </NavigationContainer>
   )
